Extract device-state-monitor pure helpers into helpers.py
Second PR in the device-state-monitor split arc. Mechanical-only:

Move the module-level pure functions (no class state, no
callbacks) out of ``_device_state_monitor/controller.py`` into a
new ``helpers.py`` sibling module:

* ``_normalize_mac`` (+ its ``_MAC_SEPARATORS`` constant)
* ``_http_url_from_service_info``
* ``_decode_txt_bytes_to_sorted_pairs``
* ``_decode_mdns_txt_records``
* ``device_name_from_service`` (public)
* ``_pick_ipv4``

``_ESPHOME_SERVICE_TYPE`` constant moves with them since
``_decode_txt_bytes_to_sorted_pairs`` uses it as a placeholder
string for the ``AsyncServiceInfo`` ctor; controller.py imports
it back from helpers so the existing 8 usages keep the bare name.

Public name (``helpers.py``, no underscore) matches the
convention from ``controllers/firmware/helpers.py`` and
``controllers/devices/helpers.py``.

``__init__.py`` re-exports ``_decode_txt_bytes_to_sorted_pairs``
and ``device_name_from_service`` source-redirected from
``.helpers`` (was ``.controller``). External callers and tests
keep working unchanged.

controller.py: 1638 → 1453 lines.
